+++ images/ubuntu/scripts/build/install-miniconda.sh
@@ -8,14 +8,18 @@
 source $HELPER_SCRIPTS/etc-environment.sh
 
 # Install Miniconda
-curl -fsSL https://repo.anaconda.com/miniconda/Miniconda3-latest-Linux-x86_64.sh -o miniconda.sh \
-    && chmod +x miniconda.sh \
-    && ./miniconda.sh -b -p /usr/share/miniconda \
-    && rm miniconda.sh
+if [ ! -d /usr/share/miniconda ]; then
+    curl -fsSL https://repo.anaconda.com/miniconda/Miniconda3-latest-Linux-x86_64.sh -o miniconda.sh \
+        && chmod +x miniconda.sh \
+        && ./miniconda.sh -b -p /usr/share/miniconda \
+        && rm miniconda.sh
+else
+    echo "Miniconda already installed, skipping."
+fi
 
 CONDA=/usr/share/miniconda
 set_etc_environment_variable "CONDA" "${CONDA}"
 
-ln -s $CONDA/bin/conda /usr/bin/conda
+ln -sfn $CONDA/bin/conda /usr/bin/conda
 
 invoke_tests "Tools" "Conda"
